Janitor QOL (#17594)
* Yes

* No

suggested changes

* Maybe

Suggestions + water in cart sprites

* Merge Conflict Resolve

* Suggested Changes :o)

* fixes bugs

* suggested changes round 2

+++ code/modules/reagents/reagent_containers/glass_containers.dm
@@ -283,7 +283,6 @@
 /obj/item/reagent_containers/glass/beaker/drugs/meth
 	list_reagents = list("methamphetamine" = 10)
 
-
 /obj/item/reagent_containers/glass/bucket
 	desc = "It's a bucket."
 	name = "bucket"
@@ -317,6 +316,10 @@
         reagents.clear_reagents()
 
 /obj/item/reagent_containers/glass/bucket/attackby(obj/D, mob/user, params)
+	if(istype(D, /obj/item/mop))
+		var/obj/item/mop/m = D
+		m.wet_mop(src, user)
+		return
 	if(isprox(D))
 		to_chat(user, "You add [D] to [src].")
 		qdel(D)
